requirements.txt
streamlit
google-genai
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
faiss-cpu
sentence-transformers
pdf2image
python-dotenv
Pillow
sqlalchemy
pytesseract
boto3